Fix randomly added stuff after rename :skull:
