+++ src/main/java/ca/mcmaster/se2aa4/island/team037/algorithm/mvp/StopState.java
@@ -1,9 +1,9 @@
-package ca.mcmaster.se2aa4.island.teamXXX.algorithm.mvp;
+package ca.mcmaster.se2aa4.island.team037.algorithm.mvp;
 
-import ca.mcmaster.se2aa4.island.teamXXX.actions.Action;
-import ca.mcmaster.se2aa4.island.teamXXX.algorithm.State;
-import ca.mcmaster.se2aa4.island.teamXXX.drone.Drone;
-import ca.mcmaster.se2aa4.island.teamXXX.result.ActionResult;
+import ca.mcmaster.se2aa4.island.team037.actions.Action;
+import ca.mcmaster.se2aa4.island.team037.algorithm.State;
+import ca.mcmaster.se2aa4.island.team037.drone.Drone;
+import ca.mcmaster.se2aa4.island.team037.result.ActionResult;
 
 public class StopState extends State {
