+++ src/main/java/ca/mcmaster/se2aa4/island/team037/algorithm/spiral/SpiralSearchDroneAlgorithm.java
@@ -1,8 +1,8 @@
-package ca.mcmaster.se2aa4.island.teamXXX.algorithm.spiral;
+package ca.mcmaster.se2aa4.island.team037.algorithm.spiral;
 
-import ca.mcmaster.se2aa4.island.teamXXX.algorithm.DroneAlgorithm;
-import ca.mcmaster.se2aa4.island.teamXXX.algorithm.State;
-import ca.mcmaster.se2aa4.island.teamXXX.drone.Drone;
+import ca.mcmaster.se2aa4.island.team037.algorithm.DroneAlgorithm;
+import ca.mcmaster.se2aa4.island.team037.algorithm.State;
+import ca.mcmaster.se2aa4.island.team037.drone.Drone;
 
 public class SpiralSearchDroneAlgorithm extends DroneAlgorithm {
