+++ src/test/java/ca/mcmaster/se2aa4/island/team037/DimensionsTest.java
@@ -1,4 +1,4 @@
-package ca.mcmaster.se2aa4.island.teamXXX;
+package ca.mcmaster.se2aa4.island.team037;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
