+++ src/test/java/ca/mcmaster/se2aa4/island/team037/result/ScanActionResultFactory.java
@@ -1,4 +1,4 @@
-package ca.mcmaster.se2aa4.island.teamXXX.result;
+package ca.mcmaster.se2aa4.island.team037.result;
 
 import java.util.Arrays;
 import java.util.List;
